Move fine-tuning MLM pipeline
